*^ CLI development fixes and updates
^ guppi/reader getNumberOfSteps, getTotalBandwidth
%^ guppi/reader getUnixDateOfLastReadBlock
^ bfr5/reader getReferencePosition

With the introduction of the Filterbank writer, the FileWriter pipeline was made generic enough to encapsulate both it and the GuppiWriter. At this point the name was changed too, seeing as it really only is an accumulation pipeline with a single module. Its accumulate function is extended...
@ pipelines/~file_writer~accumulator

++ apps/cli/ata infer Filterbank or GUPPI output from --output-type

BFR5 calibration-coefficients are no longer upchannelized by pre-beamformer channelization rate, rather the phasor frequency-index is integer-divided by the pre-beamformer channelization rate.
Furthermore, the phasor internally increments the frequency-band if it ascertains that the config.coefficients frequency dimension is larger than that of the coarse-channel equivalent of the input (ie the phasor is being used in beamforming that is incrementally processing coarse subbands instead of the whole coarse band)
@ Phasor~~Calibrations~~Coefficients
@ bfr5/reader get~Total~Dims, Boresight -> PhaseCenter
** bfr5/reader getAntennaCoefficients
*+ phasor/generic.config preBeamformerChannelizerRate

#^ radiointerferometryC99, bfr5c99